requirement.txt
gradio==4.16.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
plotly==5.18.0
scipy==1.10.1
opencv-python==4.8.0.76
pyautogui==0.9.53